aider: Remove the command to install fish plugins from the Dockerfile and add a postCreateCommand to install the fish plugins in the devcontainer.json file.
# Aider chat conversation:

USER: Fisher may not be available at that time, as it's provided as a devcontainer feature. Maybe this would be better as a postCreateCommand

ASSISTANT: To install the `PatrickF1/fzf.fish` plugin using the fisher plugin manager as a postCreateCommand, we need to modify `.devcontainer/devcontainer.json`:

1. Remove the command to install fish plugins from the Dockerfile.
2. Add a postCreateCommand to the devcontainer.json file to install the fish plugins.

Here are the *SEARCH/REPLACE* blocks:

.devcontainer/Dockerfile
```Dockerfile
<<<<<<< SEARCH
# Install fish plugins
RUN fish -c "fisher install PatrickF1/fzf.fish"
=======
>>>>>>> REPLACE
```

.devcontainer/devcontainer.json
```json
<<<<<<< SEARCH
"postCreateCommand": "poetry install --no-root"
=======
"postCreateCommand": [
    "poetry install --no-root",
    "fish -c 'fisher install PatrickF1/fzf.fish'"
]
>>>>>>> REPLACE
```

This change will remove the line `RUN fish -c "fisher install PatrickF1/fzf.fish"` from the Dockerfile and add the command `fish -c 'fisher install PatrickF1/fzf.fish'` to the postCreateCommand array in the devcontainer.json file. This will ensure that the fisher plugin manager is available when the command is run.