aider: Add 'poetry shell' command to automatically activate the virtual environment.
# Aider chat conversation:

USER: Does it make sense to add the 'poetry shell' command as well?

ASSISTANT: The `poetry shell` command is used to spawn a shell within the context of the Python virtual environment. If you want to automatically activate the virtual environment every time the devcontainer is created, you can add the `poetry shell` command to the `postCreateCommand` section of your `.devcontainer/devcontainer.json` file.

However, please note that the `poetry shell` command is blocking, meaning it will keep the terminal session open and won't return control to the terminal until you exit the shell. This might interfere with other commands you want to run after it in the `postCreateCommand` section.

If you still want to add it, here is the *SEARCH/REPLACE* block:

.devcontainer/devcontainer.json
```json
<<<<<<< SEARCH
"postCreateCommand": [
    "poetry env use python",
    "poetry install --no-root"
]
=======
"postCreateCommand": [
    "poetry env use python",
    "poetry install --no-root",
    "poetry shell"
]
>>>>>>> REPLACE
```

This change will activate the virtual environment after it's created and the project's dependencies are installed.