aider: Remove Poetry feature from devcontainer.json.
# Aider chat conversation:

USER: Should it also be removed from the devcontainer.json as a feature specification then as well?

ASSISTANT: Yes, if we are installing Poetry directly in the Dockerfile, we no longer need to specify it as a feature in the `devcontainer.json` file.

Here is the *SEARCH/REPLACE* block to remove the Poetry feature from the `devcontainer.json` file:

.devcontainer/devcontainer.json
```json
<<<<<<< SEARCH
"ghcr.io/devcontainers-contrib/features/poetry:2": {},
=======
>>>>>>> REPLACE
```

This change will remove the Poetry feature from the `devcontainer.json` file.