aider: Add poetry bin path to fish_user_paths in the Dockerfile.
# Aider chat conversation:

USER: On initially opening up a default fish shell session in this devcontainer, the poetry bin path has not been added to my path, despite it being directed by a postCreateCommand. What could be the problem here?

ASSISTANT: The issue might be due to the fact that the `postCreateCommand` in the `.devcontainer/devcontainer.json` file is executed in a non-interactive shell, which means that the shell session is not persisted. Therefore, the changes made to the `fish_user_paths` variable in the `postCreateCommand` are not available in the new shell sessions.

To resolve this issue, we can add the command to update the `fish_user_paths` variable in the Dockerfile, which will be executed every time a new shell session is started.

Here is the *SEARCH/REPLACE* block:

.devcontainer/Dockerfile
```dockerfile
<<<<<<< SEARCH
# Copy the rest of your project
COPY . /app
=======
# Copy the rest of your project
COPY . /app

# Add poetry bin path to fish_user_paths
RUN echo 'set -U fish_user_paths (poetry env info --path)/bin $fish_user_paths' >> ~/.config/fish/config.fish
>>>>>>> REPLACE
```

This change will append the command to the `config.fish` file, which is executed every time a new fish shell session is started. This ensures that the poetry bin path is added to the `fish_user_paths` variable in every new shell session.